requirements_jax.txt
# Requirements for Heterodyne Detector
# Install with: pip install -r requirements_jax.txt

# CRITICAL: SDR Communication Library
pyadi-iio>=0.0.17

# Signal processing
numpy>=1.21.0
scipy>=1.7.0

# ============================================================
# GPU ACCELERATION - JAX (RECOMMENDED FOR AMD)
# ============================================================
# JAX is MUCH easier to install than CuPy for AMD GPUs!
#
# Option 1 - AMD GPU with ROCm (RX 6700 XT):
#   pip install jax[rocm] -f https://storage.googleapis.com/jax-releases/jax_rocm_releases.html
#
# Option 2 - CPU-only (still fast with JIT):
#   pip install jax jaxlib
#
# JAX benefits:
#   - Easier installation than CuPy
#   - Official AMD/ROCm support
#   - JIT compilation (fast even on CPU)
#   - Better for ML if you add it later
#   - Auto-falls back to NumPy if not installed
# ============================================================

# Uncomment ONE of these JAX options:

# For AMD GPU with ROCm:
# jax[rocm]>=0.4.20

# For CPU-only (still faster than NumPy):
# jax>=0.4.20
# jaxlib>=0.4.20

# Pattern matching
dtaidistance>=2.3.10  # DTW implementation with C speedups
python-Levenshtein>=0.20.0  # Fast Levenshtein distance

# Visualization
matplotlib>=3.5.0

# Parallel processing
joblib>=1.1.0

# Data handling
pandas>=1.3.0

# ============================================================
# INSTALLATION INSTRUCTIONS
# ============================================================
#
# 1. Basic install (CPU-only, no GPU):
#    pip install numpy scipy matplotlib pyadi-iio dtaidistance python-Levenshtein
#
# 2. With JAX GPU acceleration (AMD RX 6700 XT):
#    pip install numpy scipy matplotlib pyadi-iio dtaidistance python-Levenshtein
#    pip install jax[rocm] -f https://storage.googleapis.com/jax-releases/jax_rocm_releases.html
#
# 3. With JAX CPU acceleration (fast without GPU):
#    pip install numpy scipy matplotlib pyadi-iio dtaidistance python-Levenshtein
#    pip install jax jaxlib
#
# 4. Test JAX installation:
#    python -c "import jax; print(jax.devices())"
#
# 5. Benchmark performance:
#    python jax_pattern_matcher.py
#
# ============================================================
